split out session
